refactor observe specs helpers and keep readability gates green
